pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "motor_defects_classification"
version = "0.1.0"
description = "Awesome `motor_defects_classification` project!"
readme = "README.md"
authors = ["lsd-maddrive <letisaudep@gmail.com>"]
repository = ""
homepage = ""

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = []  #! Update me

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [  #! Update me
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
]



[tool.poetry.dependencies]
python = "~3.8.0"
typer = {extras = ["all"], version = "^0.3.2"}
rich = "^10.7.0"
tqdm = "^4.64.0"
hydra-core = "^1.2.0"
torch = "^1.12.0"
numpy = "^1.23.1"

[tool.poetry.dev-dependencies]
black = {version = "^21.7b0", allow-prereleases = true}
isort = {extras = ["colors"], version = "^^5.10.1"}

pre-commit = "^2.14.0"

nbdime = "^3.1.0"
envyaml = "^1.9.210927"

ipykernel = "^6.4.2"
jupyter = "^1.0.0"
jupyter_contrib_nbextensions = "^0.5.1"
jupyter-client = "==6.1.12"
nbconvert = "~=5.6.1"


[tool.black]
# https://github.com/psf/black
target-version = ["py38"]
line-length = 106
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 38
line_length = 106

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true



[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
  "--doctest-modules",
  "--doctest-continue-on-failure",
]
